requirements.txt
numpy==1.26.4
matplotlib==3.9.2
open-clip-torch==2.17.1
transformers==4.28.1
av==10.0.0
Pillow==9.4.0
imageio==2.27.0
imageio-ffmpeg==0.4.8
scikit-image==0.20.0
open3d==0.18.0
mmcv==1.6.0
lpips==0.1.4
diffusers==0.25.1
huggingface_hub==0.23.4
pytorch-lightning==2.4.0
xformers==0.0.28.post1
accelerate==0.23.0
plyfile
roma
tqdm
opencv-python
scipy
einops
gdown
trimesh
pyglet<2
huggingface-hub[torch]>=0.22
evo
kornia
# for sam2, customized version
-e lib/viewcrafter/extern/monst3r/third_party/sam2